Check if user omitted filetype in the filename
If the file can't be found, now we add ".bsp" to the end of the filename to see if that exists.

I worry that this might not be ideal. My thought process being that if the user omits the filetype and they have, say .bsp.uncompressed (side effect of packing via CompilePal), the program would just error out.
Unfortunately because of the fact that source mapping gives you like 5 different files with the same name but different filetypes.
Maybe down the line it'd be worth checking every file in the folder with the input filename to see if any of them are bsp files in disguise.
Potentially could be worth giving the user either an error or a prompt to choose if multiple are bsp files (ex. user has .bsp and .bsp.unpacked in the same folder)